feat: use Virtuoso directly at MarketList
