+++ src/components/MarketList/MarketList.tsx
@@ -1,17 +1,20 @@
-import { useCallback } from 'react';
+import { useCallback, useEffect } from 'react';
 import { Virtuoso } from 'react-virtuoso';
 
 import cn from 'classnames';
 import type { Market } from 'models/market';
 
-import Footer from 'components/Footer';
+import { InfoIcon } from 'assets/icons';
+
 import PredictionCard from 'components/PredictionCard';
 
-type MarketListProps = {
-  markets: Market[];
-};
+import useMarkets from 'hooks/useMarkets';
 
-export default function MarketList({ markets }: MarketListProps) {
+import { Button } from '../Button';
+import Text from '../Text';
+
+export default function MarketList() {
+  const markets = useMarkets();
   const handleItemContent = useCallback(
     (index: number, data: Market) => (
       <PredictionCard
@@ -24,14 +27,66 @@ export default function MarketList({ markets }: MarketListProps) {
     [markets.length]
   );
 
+  useEffect(() => {
+    markets.fetch();
+    // eslint-disable-next-line react-hooks/exhaustive-deps
+  }, []);
+
   return (
-    <Virtuoso
-      useWindowScroll
-      data={markets}
-      itemContent={handleItemContent}
-      components={{
-        Footer
-      }}
-    />
+    <div className="pm-c-market-list p-grid">
+      {
+        {
+          loading: (
+            <div className="pm-c-market-list__empty-state">
+              <div className="pm-c-market-list__empty-state__body">
+                <span className="spinner--primary" />
+              </div>
+            </div>
+          ),
+          error: (
+            <div className="pm-c-market-list__error">
+              <div className="pm-c-market-list__error__body">
+                <InfoIcon />
+                <Text
+                  as="p"
+                  scale="tiny"
+                  fontWeight="semibold"
+                  className="pm-c-market-list__empty-state__body-description"
+                >
+                  Error fetching markets
+                </Text>
+              </div>
+              <div className="pm-c-market-list__error__actions">
+                <Button color="primary" size="sm" onClick={markets.fetch}>
+                  Try again
+                </Button>
+              </div>
+            </div>
+          ),
+          warning: (
+            <div className="pm-c-market-list__empty-state">
+              <div className="pm-c-market-list__empty-state__body">
+                <InfoIcon />
+                <Text
+                  as="p"
+                  scale="tiny"
+                  fontWeight="semibold"
+                  className="pm-c-market-list__empty-state__body-description"
+                >
+                  There are no available markets for this category.
+                </Text>
+              </div>
+            </div>
+          ),
+          success: (
+            <Virtuoso
+              useWindowScroll
+              data={markets}
+              itemContent={handleItemContent}
+            />
+          )
+        }[markets.state]
+      }
+    </div>
   );
 }
